requirements.txt
mysql.connector
requests
